GEODE-8711: unused method cleanup, long argument, readme tweak

+++ geode-redis/src/main/java/org/apache/geode/redis/internal/ParameterRequirements/SlowlogParameterRequirements.java
@@ -53,7 +53,7 @@ public class SlowlogParameterRequirements implements ParameterRequirements {
           String.format(ERROR_UNKNOWN_SLOWLOG_SUBCOMMAND, command.getStringKey()));
     }
     try {
-      Integer.parseInt(new String(command.getProcessedCommand().get(2)));
+      Long.parseLong(new String(command.getProcessedCommand().get(2)));
     } catch (NumberFormatException nex) {
       throw new RedisParametersMismatchException(ERROR_NOT_INTEGER);
     }
